move traverseDescriptors to utils
